convert x64 assembler files to nasm syntax
files were programmatically converted using [0], with minimal manual
corrections in syscall.s.

sabotage linux[1] has a strict "no binaries" policy, so it is not possible
to bootstrap fasm, which is written in fasm, from source.

nasm otoh is written in C, and readily available as it is required as a
prerequisite for other packages.
additionally its syntax is almost identical.

[0]: http://bogdro.evai.pl/inne/fasm2nasm.txt
[1]: https://github.com/sabotage-linux/sabotage